Refactor state management into dedicated modules
